Engine: distinct VideoPlayer class that renders only to a bitmap
