+++ Engine/media/video/video.h
@@ -12,18 +12,11 @@
 //
 //=============================================================================
 //
-// Video playback interface.
-//
-// TODO: good future changes:
-//  - do not render to the screen right inside the VideoPlayer class,
-//    instead write the frame into the bitmap or texture, and expose
-//    current frame in the interface to let the engine decide what to do
-//    with it.
+// Game-blocking video interface.
 //
 //=============================================================================
 #ifndef __AGS_EE_MEDIA__VIDEO_H
 #define __AGS_EE_MEDIA__VIDEO_H
-#include <atomic>
 #include "media/audio/openalsource.h"
 #include "util/geometry.h"
 #include "util/string.h"
@@ -55,85 +48,6 @@ enum VideoSkipType
     VideoSkipKeyOrMouse   = 3
 };
 
-class IDriverDependantBitmap;
-using AGS::Common::Bitmap;
-using AGS::Common::String;
-
-// Parent video player class, provides basic playback logic,
-// queries audio and video frames from decoders, plays audio chunks,
-// renders the video frames on screen, according to the stretch flags.
-// Relies on frame decoding being implemented in derived classes.
-class VideoPlayer
-{
-public:
-    virtual ~VideoPlayer();
-    // Tries to open a video file of a given name
-    Common::HError Open(const Common::String &name, int flags);
-    virtual bool IsValid() { return false; }
-    // Stops the playback, releasing any resources
-    void Close();
-    // Begins or resumes playback
-    void Play();
-    // Pauses playback
-    void Pause();
-    // Restores the video after display switch
-    virtual void Restore() {};
-
-    // Get suggested video framerate (frames per second)
-    uint32_t GetFramerate() const { return _frameRate; }
-    // Tells if video playback is looping
-    bool IsLooping() const { return _loop; }
-    // Get current playback state
-    PlaybackState GetPlayState() const { return _playState; }
-
-    // Updates the video playback, renders next frame
-    bool Poll();
-
-protected:
-    // Opens the video, implementation-specific; allows to modify flags
-    virtual Common::HError OpenImpl(const String& /*name*/, int& /*flags*/)
-        { return new Common::Error("Internal error: operation not implemented"); };
-    // Closes the video, implementation-specific
-    virtual void CloseImpl() {};
-    // Retrieves next video frame, implementation-specific
-    virtual bool NextFrame() { return false; };
-
-    int GetAudioPos(); // in ms
-
-    int _audioChannels = 0;
-    int _audioFreq = 0;
-    SDL_AudioFormat _audioFormat = 0;
-    SoundBuffer _audioFrame{};
-    bool _wantAudio = false;
-
-    std::unique_ptr<Bitmap> _videoFrame;
-    int _frameDepth = 0; // bits per pixel
-    Size _frameSize{};
-    uint32_t _frameRate = 0u;
-
-private:
-    // Renders the current audio data
-    bool RenderAudio();
-    // Renders the current video frame
-    bool RenderVideo();
-    // Resumes after pausing
-    void Resume();
-
-    // Parameters
-    bool _loop = false;
-    int _flags = 0;
-    // Playback state
-    uint32_t _frameTime = 0u; // frame duration in ms
-    PlaybackState _playState = PlayStateInitial;
-    // Audio
-    std::unique_ptr<OpenAlSource> _audioOut;
-    // Video
-    Rect _dstRect{};
-    std::unique_ptr<Bitmap> _hicolBuf;
-    std::unique_ptr<Bitmap> _targetBitmap;
-    IDriverDependantBitmap *_videoDDB = nullptr;
-};
-
 } // namespace Engine
 } // namespace AGS
 
@@ -145,8 +59,6 @@ AGS::Common::HError play_flc_video(int numb, int flags, AGS::Engine::VideoSkipTy
 void video_pause();
 // Resume the active video
 void video_resume();
-// Update video playback if the display mode has changed
-void video_on_gfxmode_changed();
 // Stop current playback and dispose all video resource
 void video_shutdown();
