+++ Solutions/Engine.App/Engine.App.vcxproj.filters
@@ -807,6 +807,15 @@
     <ClCompile Include="..\..\Engine\media\audio\audioplayer.cpp">
       <Filter>Source Files\media\audio</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Engine\media\video\videoplayer.cpp">
+      <Filter>Source Files\media\video</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\Engine\media\video\flic_player.cpp">
+      <Filter>Source Files\media\video</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\Engine\media\video\theora_player.cpp">
+      <Filter>Source Files\media\video</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Engine\ac\asset_helper.h">
@@ -1624,6 +1633,15 @@
     </ClInclude>
     <ClInclude Include="..\..\Engine\media\audio\audioplayer.h">
       <Filter>Header Files\media\audio</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Engine\media\video\videoplayer.h">
+      <Filter>Header Files\media\video</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Engine\media\video\flic_player.h">
+      <Filter>Header Files\media\video</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Engine\media\video\theora_player.h">
+      <Filter>Header Files\media\video</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
